feat: segment descriptor of gdt

+++ gdt.h
@@ -17,7 +17,7 @@ public:
         uint8_t base_vhi;
 
     public:
-        SegmentDescriptor(uint32_t base, uint32_t limit, uint16_t type);
+        SegmentDescriptor(uint32_t base, uint32_t limit, uint8_t type);
         uint32_t Base();
         uint32_t Limit();
     } __attribute__((packed));
